requirements.txt
ultralytics>=8.0.0
flask>=2.0.0
opencv-python>=4.5.0
matplotlib>=3.5.0
numpy>=1.20.0
tqdm>=4.60.0
PyYAML>=6.0
